scene loader bug fix

+++ src/Common.cpp
@@ -28,7 +28,7 @@ namespace LTFP
     {
         Vector3r result;
 
-        for (int i = 0; i < 3; i++)
+        for (size_t i = 0; i < 3; i++)
             result[i] = computePoly(x[i], coeff[i]);
 
         return result;
